requirements.txt
agent-framework --pre
python-dotenv
azure-identity
mcp
msal
httpx
